docs(v0.24): ADR-39 立 + BACKLOG §10 结账 + 路线图 + 向量索引字节序修复 · W119
文档结账：
- 新增 docs/ADR-39-stdlib包内资源.md：stdlib 作为包内资源、jikuai.resources
  唯一定位入口、JIKUAI_STDLIB 覆盖口；推翻 ADR-16 §3.4 的 data-files 裁决
  （data_files 装到 sys.prefix，与从 site-packages/jikuai/ 上溯定位对不上，
  这才是它三个版本没落地的真实原因）；并说明为何拒 importlib.resources
  （消费方要真实文件系统路径）。
- ADR-16 §3.4 加推翻块指向 ADR-39。
- BACKLOG §10 结账，并更正「7 处定位点」清单：原文把 blocks_cli.py:186
  误算进来（那条指向 tools/），漏了 bench_compress.py:49。
- 新增 docs/路线图-v0.24.md（M30-M32 / W113-W120，§六有实施结果与三处偏离）。
- README：安装段改 pip install jikuai + 从源码开发子段、LSP 快速上手补
  pip install jikuai-lsp、项目结构树把 stdlib/ 移到 src/jikuai/ 下并把块数
  108 更正为实测 112。
- CHANGELOG 补 W119 与回归/门禁数字。

顺带修一个查证时撞见的仓库级缺陷（不是本版引入）：向量索引.bin 字节序不自洽。
文件头一直是显式小端（struct('<...')），而向量载荷写侧用 numpy .tobytes()、
读侧用 array.array('h').frombytes()——都是原生字节序。仓库里那份 bin 生成于
x86，而发的是 py3-none-any wheel：装到大端平台（s390x 等）时 int16 会被逐字节
翻转，不抛异常，只是余弦打分全错、静默返回错误的检索结果。

格式口径现定为「全小端」：
- 读侧 sys.byteorder == 'big' 时补 array.byteswap()。
- 写侧改 astype('<i2').tobytes()——在小端机器上是纯 no-op，产出字节与改前逐
  字节一致，不触碰 G12 哈希门禁、不需重跑 embeddings（本机离线跑不了）。
- 测试夹具 _write_test_index 改 struct.pack('<%dh') 显式小端，否则用例在大端
  平台上会假红。
- 新增三条：小端解析、monkeypatch 强制大端分支证明 byteswap 真被调用（防
  「守卫绿≠守卫在守」）、真实 bin 仍可加载 112 块。
真机大端验证挂在 W120。

验证：全量回归 2595 passed / 88 skipped（exit 0）；G10-G19 exit 0；
G20 单独跑 430 条条目全绿。

+++ tools/ai-bridge/generate_embeddings.py
@@ -134,7 +134,14 @@ def _write_index(path: str, names: List[str], quantized: 'numpy.ndarray',
             name_bytes = name.encode('utf-8')
             f.write(struct.pack('<H', len(name_bytes)))
             f.write(name_bytes)
-            f.write(quantized[i].tobytes())
+            # W119 · v0.24.0：此前 `quantized[i].tobytes()` 用 numpy 的**原生字节序**
+            # 写出 int16 载荷，而文件头一直用显式小端（'<HH' / '<I' / '<ff' / '<H'）——
+            # 格式自己跟自己不一致。发布走 `py3-none-any` wheel，装到大端平台（s390x 等）
+            # 上读取时 int16 会被逐字节翻转，**不抛异常，只是余弦相似度全部算错**。
+            # 格式口径现定为「全小端」，写侧显式转成小端。`astype('<i2')` 在小端机器上是
+            # 纯 no-op（字节序已经是小端），产出的字节与改动前逐字节完全一致，因此不会
+            # 触碰 G12 向量索引哈希门禁、也不需要重跑 embeddings。
+            f.write(quantized[i].astype('<i2').tobytes())
     size_kb = os.path.getsize(path) / 1024
     print(f'写入 {path} ({size_kb:.1f} KB, {count} 块, {dim} 维)')
